requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
django-filer==1.7.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-summernote==0.8.11.6
django-taggit==1.3.0
easy-thumbnails==2.7
Pillow==7.1.2
protobuf==3.12.1
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
